// alwin-web-grapescode/src/app/issues/models/issuesRequestModels.ts
export enum GetActivitiesIssueFilterEnum {
    startCreationDate = 'startCreationDate',
    endCreationDate = 'endCreationDate',
    remark = 'remark',
    operatorId = 'operatorId',
    type = 'type',
    state = 'state',
    companyId = 'companyId',
    extCompanyId = 'extCompanyId',
    issueType = 'issueType'
}